EPMLABSBRN-306 move class
